requirements.txt
pandas
numpy
scikit-learn
tensorflow
matplotlib
streamlit
pymongo
dnspython
bcrypt
python-dotenv
joblib